requirements.txt
streamlit
PyMuPDF
langchain
sentence-transformers
faiss-cpu
google-generativeai